Move function to a separate file
