+++ apis/comms_api/core/test/test_events.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import patch, AsyncMock
 
-from comms_api.core.events import create_stateful_events, send_stateless_events, parse_agent_metadata
+from comms_api.core.events import create_stateful_events, send_stateless_events
 from comms_api.models.events import StatefulEvents, StatelessEvents
 from wazuh.core.engine.models.events import StatelessEvent, Event, WazuhLocation
 from wazuh.core.indexer import Indexer
@@ -52,36 +52,3 @@ async def test_create_stateful_events(create_indexer_mock):
     create_indexer_mock.assert_called_once()
     create_indexer_mock.return_value.events.create.assert_called_once()
     assert result == expected
-
-
-@pytest.mark.parametrize('agent_id, user_agent, agent_groups, expected', [
-    (
-        'ac5f7bed-363a-4095-bc19-5c1ebffd1be0',
-        'test endpoint v5.0.0',
-        '',
-        AgentMetadata(
-            id='ac5f7bed-363a-4095-bc19-5c1ebffd1be0',
-            groups=[],
-            name='test',
-            type='endpoint',
-            version='v5.0.0'
-        )
-    ),
-    (
-        'ac5f7bed-363a-4095-bc19-5c1ebffd1be0',
-        'test2 lambda v5.0.1',
-        'default,group1,group2',
-        AgentMetadata(
-            id='ac5f7bed-363a-4095-bc19-5c1ebffd1be0',
-            groups=['default', 'group1', 'group2'],
-            name='test2',
-            type='lambda',
-            version='v5.0.1'
-        )
-    ),
-])
-def test_parse_agent_metadata(agent_id, user_agent, agent_groups, expected):
-    """Check that the `parse_agent_metadata` function works as expected."""
-    agent_metadata = parse_agent_metadata(agent_id, user_agent, agent_groups)
-
-    assert agent_metadata == expected
